Add roman numeral converter:
I meant for this to be a short exercise in using hashmaps, which turned into
me learning more about iterators, some of the cool functions that they support
("windows", "fold", "chain", etc).

There are some "bugs" with this program:

* It is not good at parsing invalid roman numerals
* It doesn't gracefully handle non roman numeral characters (just panics)

I would like to improve on this program, but I have already spent so long on it
already!