Add wordle clone:
This took way longer than I expected, but I learned a lot, and specifically
that I shouldn't use &str when I know the content will need to be dynamic,
in which case to use a String (and lifetime specifiers for simple string
objects feels like an no-no).

Also, I learned that you can just write `impl Whatever` without having to make
a trait for it first, which is nice.