Add basic command line program using structopt:
I saw a project that used this, and thought it would be fun to play around
with this. I think it is really useful for small CLI applications, but might
fall short when using large applications with more complex/conflicting
arguments. Regardless, the ability to auto-convert the args to the right type
is really nice, along with the Vec/Option support.